convert sync-events
